requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==3.12
python-dotenv
scikit-learn
scipy
dill
pyaml
mlflow
dagshub
fastapi
uvicorn




# -e .
# -e . -> this is used to trigger setup.py and it normally used
# at time of deploying any project.